unified the vnext-configure.py for mini-loop and eks deploy
