+++ packages/installer/scripts/top-level-future-install-k8s.sh
